results/rebuttal/run-20231111-034200_LLMatic-zero_shot_donnemartin-system-design-oop-eventplanner_0_20-runs_log/eventplanner/11/LLMatic/admin_dashboard.py
class AdminDashboard:
	def __init__(self):
		self.user_activities = {}
		self.system_performance = {}
		self.vendor_listings = {}

	def monitor_user_activities(self, user_id, activity):
		if user_id not in self.user_activities:
			self.user_activities[user_id] = []
		self.user_activities[user_id].append(activity)
		return self.user_activities

	def manage_system_performance(self, metric, value):
		self.system_performance[metric] = value
		return self.system_performance

	def manage_vendor_listings(self, vendor_id, listing):
		self.vendor_listings[vendor_id] = listing
		return self.vendor_listings
